Refactor BDD suite into feature and step modules
