+++ test/feature/agent_lifecycle_feature.h
@@ -1,0 +1,4 @@
+#ifndef AGENT_LIFECYCLE_FEATURE_H
+#define AGENT_LIFECYCLE_FEATURE_H
+void feature_agent_lifecycle();
+#endif
