+++ test/feature/event_publishing_feature.h
@@ -1,0 +1,4 @@
+#ifndef EVENT_PUBLISHING_FEATURE_H
+#define EVENT_PUBLISHING_FEATURE_H
+void feature_event_publishing();
+#endif
