+++ test/feature/json_serialization_feature.h
@@ -1,0 +1,4 @@
+#ifndef JSON_SERIALIZATION_FEATURE_H
+#define JSON_SERIALIZATION_FEATURE_H
+void feature_json_serialization();
+#endif
