+++ test/feature/task_scheduling_feature.h
@@ -1,0 +1,4 @@
+#ifndef TASK_SCHEDULING_FEATURE_H
+#define TASK_SCHEDULING_FEATURE_H
+void feature_task_scheduling();
+#endif
