+++ test/step/json_serialization_steps.h
@@ -1,0 +1,10 @@
+#ifndef JSON_SERIALIZATION_STEPS_H
+#define JSON_SERIALIZATION_STEPS_H
+
+namespace JsonSerializationSteps {
+void full_task_roundtrip();
+void empty_dependencies_roundtrip();
+void single_dependency_roundtrip();
+}
+
+#endif
